[ALPS02840045] mtk_cpuidle: fix coding style
[Detail]
Change coding style of saving/restoring debug registers

MTK-Commit-Id: 0898204b5d97e2f7872a7c149bcc7abc82a58e6b

CR-Id: ALPS02840045
Feature: Power Management
(cherry picked from commit 7b4510c52765a874729745a00dc6ba5907fcdded)

+++ drivers/misc/mediatek/include/mt-plat/mtk_dbg.h
@@ -17,8 +17,8 @@
 /*
 * This header is used to export functions to other driver owers.
 */
-unsigned int *mt_save_dbg_regs(unsigned int *p, unsigned int cpuid);
-void mt_restore_dbg_regs(unsigned int *p, unsigned int cpuid);
+unsigned long *mt_save_dbg_regs(unsigned long *p, unsigned int cpuid);
+void mt_restore_dbg_regs(unsigned long *p, unsigned int cpuid);
 void mt_copy_dbg_regs(int to, int from);
 
 #endif  /* ! __MT_MACH_DBG_H */
